backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pulse-backend"
version = "0.1.0"
description = "PULSE — Performance & Unit Live Scoring Engine — FastAPI backend"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.136.1",
    "uvicorn[standard]==0.46.0",
    "gunicorn==26.0.0",
    "sqlalchemy[asyncio]==2.0.49",
    "asyncpg==0.31.0",
    "alembic==1.18.4",
    "pydantic==2.13.4",
    "pydantic-settings==2.14.1",
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==5.0.0",
    "httpx==0.28.1",
    "redis[hiredis]>=5,<6",
    "openpyxl==3.1.5",
    "python-multipart>=0.0.20",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
    "ruff>=0.7",
    "mypy>=1.13",
    "pytest-cov>=5",
]
prod = []  # placeholder for prod-only adds

[project.scripts]
pulse-seed = "app.seed.__main__:main"  # wired in Plan 07; declared here for stable entrypoint

[tool.pytest.ini_options]
asyncio_mode = "auto"
# Use a single session-scoped event loop so the async SQLAlchemy engine
# (lazy-imported in conftest.db_session) stays bound to one loop across all
# tests. With function-scoped loops the engine's connection pool tries to
# re-use connections across closed loops and explodes with
# "Event loop is closed". Set both default loop scope and fixture loop scope.
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["app"]
